affinage I8 dans nuga
git-svn-id: https://elsa-svn.onera.fr/CASSIOPEE/Trunk/Cassiopee/Apps/Modules@4298 2bbd7fe0-4929-0410-bb6b-e55522daae51

+++ KCore/KCore/Nuga/include/GeomAlgo.h
@@ -22,7 +22,7 @@ template <typename ElementType>
 class GeomAlgo {
 public:
   ///
-  template <typename Coordinate_t, typename Connectivity_t, int DIM>
+  template <typename Coordinate_t, typename Connectivity_t, short DIM>
   static void neighboring (const K_FLD::ArrayAccessor<Coordinate_t>& coords,
                            const K_FLD::ArrayAccessor<Connectivity_t>& conn, Connectivity_t& neighbors);
   ///
